+++ pages/_app.js
@@ -1,10 +1,11 @@
 import { ChakraProvider } from '@chakra-ui/react'
+import '../styles/globals.css'
 
 function MyApp({ Component, pageProps }) {
   return (
-    // <ChakraProvider resetCSS={false}> 
+    <ChakraProvider>
       <Component {...pageProps} />
-    // </ChakraProvider>
+    </ChakraProvider>
   )
 }
